test: verify golden configs against Konnect
