+++ test/kongintegration/translator_golden_tests_outputs_test.go
@@ -8,7 +8,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/blang/semver/v4"
 	"github.com/go-logr/logr"
+	"github.com/kong/go-database-reconciler/pkg/dump"
 	"github.com/kong/go-database-reconciler/pkg/file"
 	"github.com/kong/go-kong/kong"
 	"github.com/samber/lo"
@@ -17,28 +19,23 @@ import (
 
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/dataplane/sendconfig"
 	"github.com/kong/kubernetes-ingress-controller/v3/test/internal/helpers"
+	"github.com/kong/kubernetes-ingress-controller/v3/test/internal/helpers/konnect"
 	"github.com/kong/kubernetes-ingress-controller/v3/test/kongintegration/containers"
 )
 
-// TestGoldenTestsOutputs ensures that the translators's golden tests outputs are accepted by Kong.
+const (
+	timeout = 5 * time.Second
+	tick    = 100 * time.Millisecond
+)
+
+// TestGoldenTestsOutputs ensures that the translators' golden tests outputs are accepted by Kong.
 func TestTranslatorsGoldenTestsOutputs(t *testing.T) {
 	t.Parallel()
-
-	const (
-		timeout = 5 * time.Second
-		tick    = 100 * time.Millisecond
-	)
-
 	ctx := context.Background()
-
-	const goldenTestsOutputsGlob = "../../internal/dataplane/translator/testdata/golden/*/*_golden.yaml"
-	goldenTestsOutputsPaths, err := filepath.Glob(goldenTestsOutputsGlob)
-	require.NoError(t, err)
-	require.NotEmpty(t, goldenTestsOutputsPaths, "no golden tests outputs found")
 
 	// TODO: Test EE features as well (requires kong/kong-gateway + license).
 	// https://github.com/Kong/kubernetes-ingress-controller/issues/4815
-	goldenTestsOutputsPaths = lo.Filter(goldenTestsOutputsPaths, func(path string, _ int) bool {
+	goldenTestsOutputsPaths := lo.Filter(allGoldenTestsOutputsPaths(t), func(path string, _ int) bool {
 		return !strings.Contains(path, "-ee/") // Skip Enterprise tests.
 	})
 
@@ -50,32 +47,6 @@ func TestTranslatorsGoldenTestsOutputs(t *testing.T) {
 	})
 
 	t.Logf("will test %d expression routes outputs and %d default ones", len(goldenTestsOutputsPaths), len(defaultOutputsPaths))
-
-	runTest := func(t *testing.T, goldenTestOutputPath string, sut sendconfig.UpdateStrategyInMemory) {
-		goldenTestOutput, err := os.ReadFile(goldenTestOutputPath)
-		require.NoError(t, err)
-
-		content := &file.Content{}
-		err = yaml.Unmarshal(goldenTestOutput, content)
-		require.NoError(t, err)
-
-		require.Eventually(t, func() bool {
-			err, resourceErrors, parseErr := sut.Update(ctx, sendconfig.ContentWithHash{Content: content})
-			if err != nil {
-				t.Logf("error: %v", err)
-				return false
-			}
-			if len(resourceErrors) > 0 {
-				t.Logf("resource errors: %v", resourceErrors)
-				return false
-			}
-			if parseErr != nil {
-				t.Logf("parse error: %v", parseErr)
-				return false
-			}
-			return true
-		}, timeout, tick)
-	}
 
 	t.Run("expressions router", func(t *testing.T) {
 		t.Parallel()
@@ -92,7 +63,7 @@ func TestTranslatorsGoldenTestsOutputs(t *testing.T) {
 
 		for _, goldenTestOutputPath := range expressionRoutesOutputsPaths {
 			t.Run(goldenTestOutputPath, func(t *testing.T) {
-				runTest(t, goldenTestOutputPath, sut)
+				ensureGoldenTestOutputIsAccepted(ctx, t, goldenTestOutputPath, sut)
 			})
 		}
 	})
@@ -112,8 +83,73 @@ func TestTranslatorsGoldenTestsOutputs(t *testing.T) {
 
 		for _, goldenTestOutputPath := range defaultOutputsPaths {
 			t.Run(goldenTestOutputPath, func(t *testing.T) {
-				runTest(t, goldenTestOutputPath, sut)
+				ensureGoldenTestOutputIsAccepted(ctx, t, goldenTestOutputPath, sut)
 			})
 		}
 	})
+}
+
+// TestGoldenTestsOutputs ensures that the translators' golden tests outputs are accepted by Konnect Control Plane
+// Admin API.
+func TestTranslatorsGoldenTestsOutputs_Konnect(t *testing.T) {
+	konnect.SkipIfMissingRequiredKonnectEnvVariables(t)
+	t.Parallel()
+
+	ctx := context.Background()
+	defaultOutputsPaths := lo.Filter(allGoldenTestsOutputsPaths(t), func(path string, _ int) bool {
+		return strings.Contains(path, "default_")
+	})
+
+	cpID := konnect.CreateTestControlPlane(ctx, t)
+	cert, key := konnect.CreateClientCertificate(ctx, t, cpID)
+	adminAPIClient := konnect.CreateKonnectAdminAPIClient(t, cpID, cert, key)
+	updateStrategy := sendconfig.NewUpdateStrategyDBModeKonnect(adminAPIClient.AdminAPIClient(), dump.Config{
+		SkipCACerts:         true,
+		KonnectControlPlane: cpID,
+	}, semver.MustParse("3.5.0"), 10)
+
+	for _, goldenTestOutputPath := range defaultOutputsPaths {
+		t.Run(goldenTestOutputPath, func(t *testing.T) {
+			ensureGoldenTestOutputIsAccepted(ctx, t, goldenTestOutputPath, updateStrategy)
+		})
+	}
+}
+
+func ensureGoldenTestOutputIsAccepted(
+	ctx context.Context,
+	t *testing.T,
+	goldenTestOutputPath string,
+	sut sendconfig.UpdateStrategy,
+) {
+	goldenTestOutput, err := os.ReadFile(goldenTestOutputPath)
+	require.NoError(t, err)
+
+	content := &file.Content{}
+	err = yaml.Unmarshal(goldenTestOutput, content)
+	require.NoError(t, err)
+
+	require.Eventually(t, func() bool {
+		err, resourceErrors, parseErr := sut.Update(ctx, sendconfig.ContentWithHash{Content: content})
+		if err != nil {
+			t.Logf("error: %v", err)
+			return false
+		}
+		if len(resourceErrors) > 0 {
+			t.Logf("resource errors: %v", resourceErrors)
+			return false
+		}
+		if parseErr != nil {
+			t.Logf("parse error: %v", parseErr)
+			return false
+		}
+		return true
+	}, timeout, tick)
+}
+
+func allGoldenTestsOutputsPaths(t *testing.T) []string {
+	const goldenTestsOutputsGlob = "../../internal/dataplane/translator/testdata/golden/*/*_golden.yaml"
+	goldenTestsOutputsPaths, err := filepath.Glob(goldenTestsOutputsGlob)
+	require.NoError(t, err)
+	require.NotEmpty(t, goldenTestsOutputsPaths, "no golden tests outputs found")
+	return goldenTestsOutputsPaths
 }
